refactor(config): restructure directory and module structure
