+++ energetica/schemas/config/level_project.py
@@ -1,7 +1,7 @@
 """LevelProjectConfig class."""
 
 from pydantic import Field
-from energetica.config.base_project_config import BaseProjectConfig
+from energetica.schemas.config.base_project import BaseProjectConfig
 
 
 class LevelProjectConfig(BaseProjectConfig):
