+++ energetica/schemas/config/storage_facility.py
@@ -2,7 +2,7 @@
 
 from pydantic import Field, RootModel
 
-from energetica.config.power_producing_facility_config import PowerProducingFacilityConfig
+from energetica.schemas.config.power_producing_facility import PowerProducingFacilityConfig
 from energetica.enums import StorageFacilityType
