+++ energetica/schemas/config/technology.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from typing import Literal
 
 from pydantic import BaseModel, Field
-from energetica.config.level_project_config import LevelProjectConfig
+from energetica.schemas.config.level_project import LevelProjectConfig
 from energetica.enums import ProjectType, TechnologyType
