+++ energetica/schemas/config/wind_power_curve.py
