+++ save_config_to_disk.py
@@ -8,13 +8,13 @@ from datetime import timedelta
 from pydantic import BaseModel
 from pydantic_yaml import to_yaml_file
 
-from energetica.config.extraction_facility_config import ExtractionFacilitiesConfig
-from energetica.config.functional_facility_config import FunctionalFacilitiesConfig
-from energetica.config.power_facility_config import PowerFacilitiesConfig
-from energetica.config.seasonal_river_discharge_config import SeasonalRiverDischargeConfig
-from energetica.config.storage_facility_config import StorageFacilitiesConfig
-from energetica.config.technology_config import TechnologiesConfig
-from energetica.config.wind_power_curve_config import WindPowerCurveConfig
+from energetica.schemas.config.extraction_facility import ExtractionFacilitiesConfig
+from energetica.schemas.config.functional_facility import FunctionalFacilitiesConfig
+from energetica.schemas.config.power_facility import PowerFacilitiesConfig
+from energetica.schemas.config.seasonal_river_discharge import SeasonalRiverDischargeConfig
+from energetica.schemas.config.storage_facility import StorageFacilitiesConfig
+from energetica.schemas.config.technology import TechnologiesConfig
+from energetica.schemas.config.wind_power_curve import WindPowerCurveConfig
 from energetica.enums import (
     ExtractionFacilityType,
     Fuel,
